Add additional notes re: ClangFormat configuration
ClangFormat has an extensive and complex configuration system. Although the documentation is excellent and
comprehensive, some additional knowledge was hard earned during the development of the Arduino ClangFormat
configuration.

Where appropriate, some of this information was recorded in comments in the original configuration file, which have
already been transferred to the notes document. However, that original comment format was limiting so some of the
information was not previously recorded there.

The additional information is hereby added to the notes document.