Sort primary keys of ClangFormat configuration
The output from the `clang-format --dump-config` command is compared against Arduino's configuration file to check for
additions, removals, or type changes resulting from updating to a new version of ClangFormat. The expectation is that
any diff represents a change in the tool which must be evaluated before updating the configuration file contents to
match the new output.

Although it seems some attempt was made at sorting the output in alphabetical order, it is not completely consistent. It
turns out that arbitrary changes to the order of the keys may occur from one version to another, resulting in spurious
diffs or difficult to read valid diffs. The solution is to sort the keys from the dumped output in a consistent order.
This is done using the excellent yq tool.

In addition to avoiding problematic diffs, this will make the configuration file easier to use.