Adjust test data for ClangFormat friendliness
The test data that targets specific ClangFormat configuration keys was designed to provide coverage with the minimum
amount of extraneous code required to make it valid C++.

While experimenting with the IndentRequires configuration key introduced in ClangFormat 13, I discovered that, although
valid code, ClangFormat does not correctly format the targeted test data for this key when the value is set to `true`.
This could be considered a sign of quality of the test data having caught unwanted output that would result from a
specific ClangFormat configuration. However, the problematic form of code, although it is appropriate for this unusual
use case of test data, is unlikely to ever occur in a real program. For this reason, I think it is best to adjust the
code slightly to a more realistic form that is also correctly handled by ClangFormat.