Update ClangFormat configuration values to 14.0.0 equivalents
The previous ClangFormat configuration was developed for use with ClangFormat 11.0.1. The version of ClangFormat now in
use by Arduino tools is 14.0.0.

Some of the configuration key type have been changed in the interim. These keys changed from a Boolean to an object data
type, but an equivalent object value is available, which allows the value to be updated without any functional change to
the output from the configuration.