Add and update unused keys in ClangFormat configuration
The previous ClangFormat configuration was developed for use with ClangFormat 11.0.1. The version of ClangFormat now in
use by Arduino tools is 14.0.0.

Configuration keys were added or expanded in the interim.

Some of the keys allows defining patterns used for matching. Arduino's configuration does not have any need to define
custom patterns. A base set of values is hard coded into ClangFormat, without any possibility of being overwritten. For
the sake of transparency, these values are written into Arduino's configuration file