Move notes re: "forced base set" ClangFormat keys to dedicated section
Some ClangFormat configuration keys have a base set of values which are hard coded into **ClangFormat**. These are
appended to any custom values set in the configuration file and thus can not be overridden.

Because this behavior of ClangFormat is unintuitive and not mentioned in the ClangFormat's documentation, notes were
written for each such key. A significant number of additional keys were added in the interim between 11.0.1 and 14.0.0,
which tipped the scales from documenting each key individually being reasonable. They are now listed under a dedicated
section of the notes instead to improve the readability and maintainability of that content.