Update value of `AllowShortIfStatementsOnASingleLine` configuration key for ClangFormat 14.0.0
The previous ClangFormat configuration was developed for use with ClangFormat 11.0.1. The version of ClangFormat now in
use by Arduino tools is 14.0.0.

The values supported for the `AllowShortIfStatementsOnASingleLine` configuration key have changed in the interim, with
the previous `Always` value no longer available. The equivalent value for ClangFormat 14.0.0 is `AllIfsAndElse`.

Since the behavior resulting from this new value is not immediately apparent from the name or the official ClangFormat
documentation, a note was added to document this.

The key value set here was chosen in order to:

- Align with behavior of the Arduino IDE 1.x "Auto Format" feature.

This was not possible to achieve using the configuration capabilities of ClangFormat 11.0.1, so the diff in the
"golden master" test data is actually a reversion of unwanted formatting changes unavoidably produced by the previous
version.