Add `BreakBeforeConceptDeclarations` key to ClangFormat configuration
The previous ClangFormat configuration was developed for use with ClangFormat 11.0.1. The version of ClangFormat now in
use by Arduino tools is 14.0.0.

This configuration key was added in the interim.

The key value set here was chosen in order to:

- Align with behavior of the Arduino IDE 1.x "Auto Format" feature.

The `false` value is the most permissive, allowing the user to break or not break at their whim.

Since the concepts feature was added to C++ relatively recently, there is not any established style in official Arduino
code.