Add `IndentAccessModifiers` key to ClangFormat configuration
The previous ClangFormat configuration was developed for use with ClangFormat 11.0.1. The version of ClangFormat now in
use by Arduino tools is 14.0.0.

This configuration key was added in the interim. The hope was that this new support for access modifier indentation
would allow Arduino's ClangFormat configuration to achieve code formatting of classes matching the Arduino code style,
and aligned with the behavior of the Arduino IDE 1.x "Auto Format" feature (which does indent them).

Unfortunately, it turns out that the setting has as an undesirable side effect on struct formatting, which renders the
indentation produced by setting it to `true` incompatible with the Arduino code style. For this reason, it was set to
`false`, which results in no functional change to the existing configuration.

A note has been added about this finding.