Add `IndentRequires` key to ClangFormat configuration
The previous ClangFormat configuration was developed for use with ClangFormat 11.0.1. The version of ClangFormat now in
use by Arduino tools is 14.0.0.

This configuration key was added in the interim.

Since the requires clauses feature was added to C++ relatively recently, there is not any established style in official
Arduino code. The key setting was chosen based on a survey of the style in use in prominent reference code.