Add `PPIndentWidth` key to ClangFormat configuration
The previous ClangFormat configuration was developed for use with ClangFormat 11.0.1. The version of ClangFormat now in
use by Arduino tools is 14.0.0.

This configuration key was added in the interim.

Since we have `IndentPPDirectives` set to `None`, this option has no effect. It is set to `-1` to cause preprocessor
directive indentation to be done according to `IndentWidth` for consistency in the case of custom user configurations
which do enable `IndentPPDirectives`.