Add `PackConstructorInitializers` key to ClangFormat configuration
The previous ClangFormat configuration was developed for use with ClangFormat 11.0.1. The version of ClangFormat now in
use by Arduino tools is 14.0.0.

This configuration key was added in the interim.

The key value set here was chosen in order to:

- Align with behavior of the Arduino IDE 1.x "Auto Format" feature.

This was not possible to achieve using the configuration capabilities of ClangFormat 11.0.1, so the diff in the
"golden master" test data is actually a reversion of unwanted formatting changes unavoidably produced by the previous
version.