Add `PenaltyIndentedWhitespace` key to ClangFormat configuration
The previous ClangFormat configuration was developed for use with ClangFormat 11.0.1. The version of ClangFormat now in
use by Arduino tools is 14.0.0.

This configuration key was added in the interim.

Since the `Penalty*` keys are not applicable when wrapping is disabled by setting `ColumnLimit: 0`, as is done in
Arduino's configuration. They are all set to an arbitrary value of `1`.