Add `ReferenceAlignment` key to ClangFormat configuration
The previous ClangFormat configuration was developed for use with ClangFormat 11.0.1. The version of ClangFormat now in
use by Arduino tools is 14.0.0.

This configuration key was added in the interim.

The key value set here was chosen in order to:

- Comply with the established Arduino code style

The chosen `Pointer` value provides consistency with the pointer alignment style.