Add `ShortNamespaceLines` key to ClangFormat configuration
The previous ClangFormat configuration was developed for use with ClangFormat 11.0.1. The version of ClangFormat now in
use by Arduino tools is 14.0.0.

This configuration key was added in the interim.

This key is only used when `FixNamespaceComments` is set to `true`. Since Arduino's configuration sets
`FixNamespaceComments: false`, the value of `ShortNamespaceLines` doesn't have any direct effect, but the chosen value
of `0` will provide the maximum possible level of consistency in custom user configurations that set
`FixNamespaceComments: true`