Add `SortJavaStaticImport` key to ClangFormat configuration
The previous ClangFormat configuration was developed for use with ClangFormat 11.0.1. The version of ClangFormat now in
use by Arduino tools is 14.0.0.

This configuration key was added in the interim.

Since this key is specific to the Java language, while Arduino's configuration is intended for use with
Arduino language/C++/C, the value of this key is irrelevant so the default value is used. This fact was documented in
the notes.