Add `SpaceAroundPointerQualifiers` key to ClangFormat configuration
The previous ClangFormat configuration was developed for use with ClangFormat 11.0.1. The version of ClangFormat now in
use by Arduino tools is 14.0.0.

This configuration key was added in the interim.

The key value set here was chosen in order to:

- Align with behavior of the Arduino IDE 1.x "Auto Format" feature.
- Comply with the established Arduino code style

The key is set to `Default` in order to preserve the behavior of emulating the Arduino IDE 1.x "Auto Format" feature's
permissive behavior by using whichever style is established in the code base (`DerivePointerAlignment: true`), or using
the Arduino code style of right alignment (`PointerAlignment: Right`) when ClangFormat can not determine an alignment
style from the code.