Add `BasedOnStyle` key to ClangFormat configuration
The ClangFormat configuration file was previously generated using ClangFormat 11.0.1. The output produced by that
version did not contain a `BasedOnStyle` key, which is why it was not present in the configuration despite being
supported by 11.0.1. Since Arduino's configuration sets all configuration keys, the `BasedOnStyle` key has no effect
anyway.

The configuration file generated using ClangFormat 14.0.0 does contain a `BasedOnStyle` key. Unfortunately, regardless
of what was set in the configuration, the tool always outputs a key set to an empty string. Even though it is likely not
an issue, this is concerning because it is not documented as a supported value.