Force valid ClangFormat `BasedOnStyle` configuration key value
In order to facilitate maintenance and development, the ClangFormat configuration file is standardized by running a
`clang-format --dump-config` command while configured according to the current configuration file, then updating the
file according to the dumped effective configuration.

The configuration file generated using ClangFormat 14.0.0 contains a `BasedOnStyle` key set to an empty string
(`BasedOnStyle: ''`), even when configured via the arbitrary `BasedOnStyle: LLVM` setting previously used by Arduino's
configuration.

Since Arduino's configuration sets all configuration keys, the `BasedOnStyle` key has no effect, and so this is likely
not an issue. However, it is concerning because an empty string is not documented as a supported value.

Out of an abundance of caution and to avoid confusion, the value of this key is reset to the documented valid value
`LLVM`.