Move collection of resources re: licence compatibility to centralized location
The subject of open source software license compatibility is complex and useful information about it difficult to find.
For this reason, a list of links to high quality references were added to the documentation for the
"Check Go Dependencies" template assets.

Equivalent templates for other project frameworks will be added. These references, which are in no way specific to the
Go modules framework, will be equally as relevant to the installation and maintenance of those assets.

Rather than duplicate the content in the documentation for each template, the list of links are hereby moved to a readme
in the folder containing the stock configuration files which are based on the information from those resources. The
documentation for the individual templates already contain a link to this readme.