Import template to validate action.yml
On every push or pull request that affects the action.yml GitHub Actions metadata file, and periodically, validate
action.yml against its JSON schema.

This is the initial import of the assets from the arduino/setup-task repository:

- https://github.com/arduino/setup-task/blob/fe5b4a500ed2a502cd93e125349a5cc854814300/.github/workflows/check-action-metadata-task.yml
- https://github.com/arduino/setup-task/blob/fe5b4a500ed2a502cd93e125349a5cc854814300/Taskfile.yml#L28-L35

They will now be hosted in this centralized location to facilitate their maintenance and use in the the Arduino Tooling
Team's other GitHub Actions actions projects.