Add contributor guide templates
Arduino tooling projects are produced through the collaboration of numerous contributors.

High quality documentation of how to contribute to the projects can give everyone the opportunity to participate, while
also reducing maintenance effort and increasing quality of contributions.

Producing such documentation for each project can seem a daunting task. Those who are knowledgeable enough to do so are
also the ones less likely to recognize the need for it. These factors have resulted in less than ideal coverage of
contribution documentation in the Arduino Tooling projects.

A collection of "template" contribution guides are established here. These are intended to be easily applied to Arduino
tooling projects, reducing the effort to establish and maintain contribution documentation.

This guides document the various ways of contributing to the project.

It takes advantage of GitHub's "contributing guidelines" feature to increase the visibility of the information to
prospective contributors. The promoted "CONTRIBUTING.md" file is used as an entry point to the guide, with links from
there leading to dedicated guides for each specific type of contribution. In this way, the contributor is only presented
with relevant information, improving the approachability and readability of the content.