Update publish-go-nightly-task using parallelization
Introducing the use of a matrix, greatly improves time performances during the build making process, since each build task is performed simultaneously.
The calculation of the checksums has also been modified. Previously, it was done three (3) times at different stages of the workflow. This is pointless, since the only checksums that matter are the ones
calculated when the files are in their final form. For this reason, it is now only done once during the release creation.