Update publish-go-tester-task using parallelization
Introducing the use of a matrix, greatly improves time performances during the build making process, since each build task is performed simultaneously. Furthermore, it allows to directly upload a different
artifact for each build. To complete this set of changes, a separate job to calculate the checksums and upload the relative artifact has been added to the workflow.