Use correct task name to build project in release workflow
The "Release" workflow (Go, Task, CGO) template was recently changed to use a job matrix to run builds for each target
in parallel. The new design is intended to use a dedicated task for each target instead of the previous approach of
executing a single `dist:all` task.

The task name was not updated at that time, which caused the workflow to fail:

task: Task "dist:all" not found

The task invocation is hereby updated to use the target-specific task names.