Reduce line length to <=120 in YAML files where feasible
120 columns is the recommended line length for YAML code in Arduino tooling projects. The yamllint tool used by the
"Check YAML" template produces a warning when a line exceeds this length.

This is not a hard limit and in some cases it is either impossible or not beneficial to make lines less than 120 in
length so some violations of the guideline are unavoidable. However, a survey of the YAML files in the repository
revealed some opportunities for improving the code by reducing the lengths.