Install npm dependencies in "Check npm Dependencies" template
The "Check npm Dependencies" template is used to check for compatibility of dependencies with the licensing of projects
that use npm for dependency management.

The "Licensed" tool is used for this check. This tool requires the dependencies to have been installed in advance.

Previously, the template's task did not install the dependencies. This caused the check to fail whenever the
dependencies were not already installed by chance.

The dependency check tasks are intended to be applicable to projects of any type since the Licensed tool supports
multiple project types. Each project type might have different requirements for advance preparation of the dependencies.
A call to a `general:install-deps` task was added to the dependency check tasks for this purpose. A different variant of
that task will be provided for each specific project type. In project types where no preparation is needed, an empty
`general:install-deps` task must be added in order to allow the task call to succeed.