Add infrastructure for checking project filesystem
There are differences in the filename restrictions between operating systems. The use of filenames that are not valid
one one operating system in the project will cause problems for contributors or users (e.g., not possible to check out
the repository).

Tasks are added to check for non-portable filenames:

- Presence of characters prohibited by an operating system in filenames
- Use of filenames reserved by an operating system

Tasks are also added to check for problems with symbolic links ("symlinks") contained in the project:

- Broken symlinks
- Circular symlinks

A GitHub Actions workflow is included to run the tasks on any change to the project files in order to avoid the
introduction of problems with the project filesystem, and periodically in order to catch breakage caused by external
changes.

Co-authored-by: Umberto Baldi <34278123+umbynos@users.noreply.github.com>