Add support to "Check npm" workflow for projects in subfolders
The "Check npm" workflow checks for problems with the npm configuration files of a repository. Previously this workflow
assumed that a repository would only ever have these files in the root folder. However, a repository might also contain
multiple separate npm-managed projects in arbitrary subfolders.

Support for any repository structure is added to the workflow by using a job matrix to check the projects under an array
of arbitrary paths that can be configured for the specific repository the template is installed in.