Use modern version of validator to validate Dependabot configuration files
The repository infrastructure validates the project's own as well as the "asset" Dependabot configuration files against
the JSON schema. The ajv-cli validator is used to perform the validation.

Support for the JSON schema "Draft-04" specification was dropped at the 4.0.0 release of ajv-cli. This means that the
previous version (3.3.0) must be used when validating files against a JSON schema based on the "Draft 4" specification.
When working with schemas based on newer specification versions, it is preferred to use the latest version of ajv-cli in
order to benefit from the ongoing development on the tool. For this reason, two different versions of ajv-cli are used.

At the time the system for validating the Dependabot configuration files was set up, that JSON schema was based on the
"Draft-04" specification, so ajv-cli 3.3.0 was used. Since that time the schema was updated to "Draft-07" so it is now
possible to use the modern ajv-cli for this validation.