Add support to "Check License" for checking license files in multiple paths
In cases where a project contains distinct components in subfolders, multiple license files might be present.

Previously the "Check License" workflow only supported checking the license file in the root of the repository. Support
for validating an arbitrary number of license files with arbitary locations, types, and filenames is added. A job matrix
is used to provide this support in a manner that makes application-specific configuration of the workflow easy and
without code duplication.