Upgrade actions/upload-artifact@v4, actions/download-artifact@v4 and geekyeggo/delete-artifact@v5
The workflow previously uploaded multiple artifacts under the same
'name'. This is no more allowed on v4, so we have to upload each
artifact under a unique name. When downloading we can merge multiple
artifacts using the 'pattern' matching and the `merge-multiple: true`
option.
The geekyeggo/delete-artifact instead can use the pattern matching
directly in the `name` field.