Configure actions/upload-artifact action to upload required hidden files
A breaking change was made in the 3.2.1 release of the "actions/upload-artifact" action, without doing a major version
bump as would be done in a responsibly maintained project. The action now defaults to not uploading "hidden" files.

The "Check Go Dependencies" and "Check npm Dependencies" templates use the "Licensed" tool to check for incompatible
dependency licenses. The dependencies license metadata cache used by Licensed is stored in a folder named `.licensed`.

In order to facilitate updates, the workflows of these templates upload the generated dependencies license metadata
cache as a workflow artifact when the current cache is found to be outdated.

The `.` at the start of the `.licensed` folder name causes it to now not be uploaded to the workflow artifact. In order
to catch such problems, the workflow configures the "actions/upload-artifact" action to fail if no files were uploaded.
So the workflow now fails:

Error: No files were found with the provided path: .licenses/. No artifacts will be uploaded.

The problem is fixed by disabling the "actions/upload-artifact" action's new behavior via the `include-hidden-files`
input. After this change, the workflows can once more upload the dependencies license metadata cache to a workflow
artifact as needed