Use environment variable to define build folder name in "Release" workflows
Multiple commands in the release workflows include the path of the folder that contains the build output. Previously,
the complex folder name was specified redundantly in each instance.

The readability and maintainability of the workflow is improved by using an environment variable to define the folder
name in a single place, then referencing that environment variable in each of the individual commands that use the path.