Remove unused job outputs from release workflows
In a previous revision of the release workflows, the updated checksums were determined by the macOS notarization job and
then passed to the subsequent job via job outputs. That approach was changed during a later refactoring, but the code
that declares the job outputs was not removed at that time.

The checksum job output declaration code is now unused (and unusable since the checksum determination and output
definitions were removed) so it only makes the workflows more difficult to understand and maintain. For this reason, the
vestigial code is hereby removed from the workflows.